requirements.txt
asgiref==3.8.1
Django==5.0.6
lxml==5.3.0
pillow==10.3.0
python-docx==1.1.2
sqlparse==0.5.0
typing_extensions==4.12.2
